Change and rename the 'escape' function interface to return the size of the escaped string, and take an optional 'char **' to store the location of the beginning of an invalid escaped sequence.
Fix a bug in the escaped strings printing operation:
using fwrite instead of fputs, since escaped string may contain '\000'
characters, and fputs stop printing on '\0' characters. But fwrite use
a size limit on writing which is easly provided by the string object
len field or the result of the 'escape' function. An example:

 **using fputs**
     print('Hello,\000World!')  # => prints Hello,

 **using fwrite**
     print('Hello,\000World!')  # => prints Hello,World!