Propagate the internal details of dealing with raw strings to the lexer and the parser, and eliminate any code dealing with raw string in the evaluator, like the 'israw' flag in the Str_obj. Now the raw strings are just like regular strings, but are stored internally as escaped strings instead:
`Hello\nThere` ==(parse)==> 'Hello\\nThere'

Rename the mistakenly named function 'strescp' to 'strunescp', since
it unescapes the strings not the opposite.

Fix a corner case in 'strunescp':
unescape the raw string `\` (internally '\\'). The bug occurs in the
string iterator, since the for loop internally emits a one char string
on each loop.

Add the function 'strescp', which "correctly" escapes a string. It's
used to escape the raw strings at the parser level.